Conditional installation of powerline fonts

+++ install.sh
@@ -90,8 +90,10 @@ echo "[+] Installing misc. packages from pip ..."
 
 
 # Install powerline fonts
-echo "[+] Installing powerline fonts ..."
-( git clone https://github.com/powerline/fonts.git --depth=1 && cd fonts && ./install.sh && cd ..  && rm -rf fonts ) &>>$LOG_FILE
+if [ $(ls ~/.local/share/fonts | grep 'Powerline' | wc -l) -lt 50 ]; then
+    echo "[+] Installing powerline fonts ..."
+    ( git clone https://github.com/powerline/fonts.git --depth=1 && cd fonts && ./install.sh && cd ..  && rm -rf fonts ) &>>$LOG_FILE
+fi
 
 # Install base16-shell and set theme
 if [ ! -d ~/.config/base16-shell ]; then
